+++ CHANGELOG.md
@@ -41,6 +41,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
    identifier.
 -  Fix detection of macro definitions that end on comment.
 -  Fix update of text document contents.
+-  Fix incremental update of abstract syntax trees.
 
 
 [0.7.1] - 2026-04-21